Assume contiguous 1D array. Input no-data policy.
All algorithms taking an array or a raster assume the values are
stored in a contiguous array. Whenever possible a single index
is used to iterate over the values. This is necessary to prevent
calculation of element positions as much as possible.

Input no-data is now read directly from the arguments. The
algorithms don't assume that the no-data is already merged
and stored in the result. Preprocessing input no-data is silly
(cache misses, additional work).